Add elm-pages script for generating CLDR files
This replaces the elm-cldr-generator repo, which contained this repo as a sub-module and used `elm-cli`.
I have found that the structure of the CLDR files changes enough from release to release that an ergonomic generation script is necessary, and elm-pages is much easier to modify than elm-cli.
This version of the generator script runs the cleanup commands itself, instead of just reminding you  to do so.